refactor arena java sdk
