+++ docs/serving/seldon-core/pre-packaged-model-server.md
@@ -1,26 +1,32 @@
 This guide walks through the steps to deploy and serve a model with seldon core pre-packaged model server.
 
-1. Setup 
-   
+1\. Setup
+
 Follow the Seldon Core [guide](https://github.com/SeldonIO/seldon-core#install-seldon-core) to install Seldon Core.For the prerequisites,you should ensure 8g memory and 4 core cpu available in your environment.
 
-2. Summit your serving job into seldon core
-```shell script
+
+2\. Summit your serving job into seldon core
+
+```shell
 $ arena serve seldon --implementation SKLEARN_SERVER --modelUri gs://seldon-models/sklearn/iris --name sklearn-iris
 configmap/sklearn-iris-202102222213-seldon-core-serving created
 configmap/sklearn-iris-202102222213-seldon-core-serving labeled
 seldondeployment.machinelearning.seldon-core.io/sklearn-iris created
 
 ```
-3. List the job you were just serving
-```shell script
+
+3\. List the job you were just serving
+
+
+```shell
 $ arena serve list 
 NAME          TYPE        VERSION       DESIRED  AVAILABLE  ADDRESS       PORTS
 sklearn-iris  Seldon      202102222213  1        1          172.29.11.50  RESTFUL:8000,GRPC:5001
 ```
-4. Test the model service
 
-```shell script
+4\. Test the model service
+
+```shell
 $ kubectl get svc
 NAME                                          TYPE        CLUSTER-IP      EXTERNAL-IP   PORT(S)             AGE
 sklearn-iris-default                          ClusterIP   172.29.2.217    <none>        8000/TCP,5001/TCP   2m8s
@@ -50,8 +56,10 @@ $ curl -X POST http://localhost:8000/api/v1.0/predictions -H 'Content-Type: appl
 }
 
 ```
-5. Delete the inference service
-```shell script
+
+5\. Delete the inference service
+
+```shell
 $ arena serve delete sklearn-iris                                                                                          
 seldondeployment.machinelearning.seldon-core.io "sklearn-iris" deleted
 configmap "sklearn-iris-202102222213-seldon-serving" deleted
