+++ sdk/arena-java-sdk/pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>com.aliyun.ack</groupId>
     <artifactId>arena-java-sdk</artifactId>
-    <version>1.0.0-SNAPSHOT</version>
+    <version>1.0.1-SNAPSHOT</version>
 
     <dependencies>
         <dependency>
@@ -17,13 +17,13 @@
         <dependency>
             <groupId>io.kubernetes</groupId>
             <artifactId>client-java</artifactId>
-            <version>6.0.1</version>
+            <version>10.0.0</version>
     		<exclusions>
         		<exclusion>
             		<groupId>io.sundr</groupId>
             		<artifactId>builder-annotations</artifactId>
         		</exclusion>
-    	</exclusions>
+    	    </exclusions>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -37,9 +37,10 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.1</version>
                 <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
+                    <source>11</source>
+                    <target>11</target>
                 </configuration>
             </plugin>
             <plugin>
@@ -58,16 +59,5 @@
             </plugin>
         </plugins>
     </build>
-
-    <distributionManagement>
-        <repository>
-            <id>releases</id>
-            <url>http://mvnrepo.alibaba-inc.com/mvn/releases</url>
-        </repository>
-        <snapshotRepository>
-            <id>snapshots</id>
-            <url>http://mvnrepo.alibaba-inc.com/mvn/snapshots</url>
-        </snapshotRepository>
-    </distributionManagement>
 
 </project>
