+++ sdk/arena-java-sdk/src/main/java/com/github/kubeflow/arena/utils/Utils.java
@@ -1,10 +1,9 @@
 package com.github.kubeflow.arena.utils;
 
-import io.kubernetes.client.ApiClient;
-import io.kubernetes.client.Configuration;
+import io.kubernetes.client.openapi.ApiClient;
+import io.kubernetes.client.openapi.Configuration;
 import io.kubernetes.client.util.KubeConfig;
 import io.kubernetes.client.util.ClientBuilder;
-import java.util.concurrent.TimeUnit;
 import java.io.FileReader;
 import java.io.IOException;
 
@@ -18,7 +17,7 @@ public class Utils {
             client = ClientBuilder.cluster().build();
         }
         Configuration.setDefaultApiClient(client);
-        return  client;
+        return client;
     }
 
 }
